feat(ui): collapsible sidebar and experimental settings popup
- Sidebar collapses via an edge handle or Cmd+Ctrl+S, with a labeled
  toggle, a persisted resizable width, and a reset action
- "Search containers" bar at the sidebar foot on non-dashboard pages,
  taking the slot where the carousel title shows on the dashboard
- Optional fullscreen settings popup (default off, experimental) that
  reuses the settings panel and opens from the topbar cog or the actions
  menu, so preferences change without leaving the page. The /settings
  route is unchanged.
