+++ assets/components/common/KeyShortcut.vue
@@ -7,6 +7,9 @@
       <ph:command v-if="isMac" class="size-4" />
       <ph:control-bold v-else class="size-4" />
     </template>
+    <template v-if="modifiers.includes('^')">
+      <ph:control-bold class="size-4" />
+    </template>
     <kbd class="kbd kbd-xs ml-0.5 uppercase">{{ char }}</kbd>
   </div>
 </template>
